Fixes
fixes are made
