+++ EMTodoList/Errors.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class CustomError: Error { // +
+class CustomError: Error {
     var message: String
     init(message: String) {
         self.message = message
